Do single-recipe runs triggered by the socket
External programs can connect to the chef control socket and trigger
full or single-recipe runs using a simple protocol.

The protocol is subject to change, but for now it is length + recipe
string. The length must be an integer packed with pack("N").

recipe = "my_cookbook::my_recipe"
s = UNIXSocket.new("/tmp/chef_socket")
s.write([recipe.size].pack("N") + recipe)

A full run can be triggered by sending a length of 0:

s.write([0].pack("N"))

If a run is triggered while a run is in progress, it will be enqueued
and run immediately after the current run is complete.